+++ dyndis/implementor.py
@@ -2,6 +2,7 @@ from functools import partial
 from itertools import chain
 from numbers import Number
 from typing import NamedTuple, Callable, List, Union
+from warnings import warn
 
 from dyndis.candidate import Candidate, Self
 
@@ -28,6 +29,7 @@ class Implementor:
     """
     A descriptor that holds candidates that include a class until it is complete
     """
+
     def __init__(self, multidispatch):
         """
         :param multidispatch: the multidispatch to add the candidates to
@@ -40,13 +42,21 @@ class Implementor:
     def __set_name__(self, owner, name):
         """
         indicates that the class is fully created, adds the queued implementors to the multidispatch
-        :param owner:
-        :param name:
-        :return:
         """
+
+        def check_candidate(candidate):
+            if self.multidispatch.__name__ is not None \
+                    and candidate.__name__ is not None\
+                    and self.multidispatch.__name__.strip('_') != candidate.__name__.strip('_'):
+                warn(f'implementor {candidate} has dissimilar name from multidispatch {self.multidispatch}')
+            return candidate
+
         if self.locked:
             return
-        self.multidispatch.add_candidates(chain.from_iterable(q.cands(owner) for q in self.queue))
+        self.multidispatch.add_candidates(
+            check_candidate(c) for c in
+            chain.from_iterable(q.cands(owner) for q in self.queue)
+        )
         self.locked = True
 
     def implementor(self, priority=0, symmetric=False, func=None) \