+++ dyndis_data/__init__.py
@@ -1,3 +1,3 @@
 __author__ = 'ben avrahami'
-__version__ = '0.0.3b0'
+__version__ = '0.0.3post'
 __url__ = "https://github.com/bentheiii/dyndis"